refactor: address Clippy lints and bundle long argument lists
- Introduce WorkerPoolConfig and ExecutionParams structs
- Extract process_item helper from worker loop body
- Simplify error mapping, match arms and Default derive
- Remove unnecessary casts and references
- Rename model from_str helpers to parse_str

+++ src/api/health.rs
@@ -402,7 +402,7 @@ async fn assess_disk(state: &AppState) -> HealthAssessment {
                     let stat = nix::sys::statvfs::statvfs(path);
                     match stat {
                         Ok(s) => {
-                            let avail_bytes = s.blocks_available() as u64 * s.block_size() as u64;
+                            let avail_bytes = s.blocks_available() * s.block_size();
                             let gb = avail_bytes as f64 / (1024.0 * 1024.0 * 1024.0);
                             if gb < 0.1 {
                                 degraded = true;
